backend/requirements.txt
# Medical Intake Summarizer - Backend Dependencies
# Core FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and serialization
pydantic==2.5.0

# HTTP client for API calls
httpx==0.25.2
requests==2.31.0

# OpenAI integration
openai==1.3.7

# Environment variable management
python-dotenv==1.0.0

# CORS middleware
python-multipart==0.0.6

# Security and authentication (for future use)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.9

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1 